+++ src/data/error.rs
@@ -1,36 +1,23 @@
-use core::fmt;
-use error_stack::Context;
+use snafu::prelude::*;
 
 /// MAX31855Data error type
-#[derive(Debug)]
+#[derive(Debug, Clone, Snafu)]
+#[snafu(visibility(pub(super)))]
 pub enum MAX31855DataError {
-    /// Internal (currently this can be only linearization) error
-    Internal,
+    #[cfg(feature = "linearization")]
+    /// Linearization error
+    #[snafu(display("linearization error"))]
+    Linearization { source: crate::LinearizationError },
     /// Unknown error
+    #[snafu(display("unknown error"))]
     Unknown,
     /// Errors when the thermocouple is open (no connections)
+    #[snafu(display("thermocouple is open (no connections)"))]
     OpenConnection,
     /// Errors when the thermocouple is short-circuited to Vcc
+    #[snafu(display("thermocouple is short circuited to Vcc"))]
     ShortVCC,
     /// Errors when the thermocouple is short-circuited to GND
+    #[snafu(display("thermocouple is short circuited to Ground"))]
     ShortGround,
 }
-
-impl fmt::Display for MAX31855DataError {
-    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt.write_str("Cannot convert MAX31855 data")?;
-        match self {
-            MAX31855DataError::Internal => Ok(()),
-            MAX31855DataError::Unknown => fmt.write_str(": unknown error"),
-            MAX31855DataError::OpenConnection => fmt.write_str(": thermocouple is open"),
-            MAX31855DataError::ShortVCC => {
-                fmt.write_str(": thermocouple is short circuited to Vcc")
-            }
-            MAX31855DataError::ShortGround => {
-                fmt.write_str(": thermocouple is short circuited to Ground")
-            }
-        }
-    }
-}
-
-impl Context for MAX31855DataError {}